grade

+++ grade.sh
@@ -1,13 +1,14 @@
-#set -e
-
 rm -rf student-submission
-rm ListExamples.java
-rm ListExamples.class
+if [ -e "student-submission/ListExamples.java" ]
+then
+	rm ListExamples.java
+	rm ListExamples.class
+fi
 git clone $1 student-submission
 
 if [ -e "student-submission/ListExamples.java" ]
 then
-	echo "ListExamples file found"
+	echo "ListExamples file found: +20 points"
 	cp student-submission/ListExamples.java ./
 	javac -cp .:lib/hamcrest-core-1.3.jar:lib/junit-4.13.2.jar *.java 2> error.txt
 	if [ -s error.txt ]
@@ -15,18 +16,24 @@ then
 		echo "Compiler error"
 		echo "$(cat error.txt)"
 		echo "Exited with error code $?"
-		exit
+		echo "Score: 20/100 for proper file submission but compile error"
+		exit 1
 	else
-		echo "Comile success"
+		echo "Comile success: +30"
 		java -cp .:lib/hamcrest-core-1.3.jar:lib/junit-4.13.2.jar org.junit.runner.JUnitCore TestListExamples > output.txt
-		if [ "$(grep -o E output.txt | wc -l)" != 0 ]
+		if grep "OK (2 tests)" $"output.txt"
 		then
+		  echo "No JUnit test errors: +25 points per passed test"
+		  echo "Score: 100/100"
+		  exit 0
+		else
 		  echo "JUnit test errors"
 		  echo "$(cat output.txt | tail -n2)"
-		else
-		  echo "No JUnit test errors"
+		  exit 1
 		fi
 	fi
 else
 	echo "ListExamples file not found"
+	echo "Score: 0/100"
+	exit
 fi
